1st Commit
I transfered it to 2D instead of 3D, now everything should be good. But before we start on the game we need to work out who is allowed to touch what files.
Make sure it actually is 2D for you when you start using Unity, I can't tell if GitHub detected the change or not.